LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/text/2d.text.font.parse.invalid-expected.txt

2d.text.font.parse.invalid


FAIL OffscreenCanvas test: 2d.text.font.parse.invalid assert_equals: ctx.font === '20px serif' (got bogus[string], expected 20px serif[string]) expected "20px serif" but got "bogus"
